Fitting Target

Fitting Target

tar_fit(
  name,
  object,
  formula,
  data,
  control = parsnip::control_parsnip(),
  ...,
  tar_args = list()
)

Arguments

name
Name of target

object
An object of class model_spec that has a chosen engine (via set_engine()).

formula
An object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.

data
Optional, depending on the interface (see Details below). A data frame containing all relevant variables (e.g. outcome(s), predictors, case weights, etc). Note: when needed, a named argument should be used.

control
A named list with elements verbosity and catch. See control_parsnip().

...
Not currently used; values passed here will be ignored. Other options required to fit the model should be passed using set_engine().

tar_args
Arguments passed to tar_target.

Value

A tar_target object.